requirements.txt
chainlit==2.3.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.83
langchain-openai==0.2.0
langchain-chroma==0.1.4
python-dotenv==1.0.1
chromadb==0.5.5
duckduckgo_search==6.2.12
pypdf>=4.0
